Make classification tasks async using Redis
Classication task is immediately queued into a Redis server.
A RQ worker is provided: it can be run on another machine or on the
same machine.
